标签展示功能，评论功能，审核，回复，删除

+++ src/main/java/com/itomelet/blog/controller/admin/BlogController.java
@@ -150,7 +150,7 @@ public class BlogController {
 
     @PostMapping("/blogs/delete")
     @ResponseBody
-    public boolean delete(@RequestBody Integer[] ids) {
+    public boolean delete(@RequestBody Long[] ids) {
         return blogService.deleteBlog(ids);
 
     }
